Adicionando um as type para tratar varáveis categoricas como inteiros ao invés de floats

+++ classifica_buscas.py
@@ -8,7 +8,8 @@ df = pd.read_csv('busca.csv')
 X_df = df[['home', 'busca', 'logado']]
 Y_df = df['comprou']
 
-Xdummies_df = pd.get_dummies(X_df)
+#as type garante que o valor vai ser do tipo inteiro
+Xdummies_df = pd.get_dummies(X_df).astype(int)
 #Y não precisa ser transformado
 Ydummies_df = Y_df
